refactor: 좋아요 버튼 피플,프로젝트에서 재사용 가능하게 수정
